Move out functionality from common for readability.
